services/user-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
email-validator==2.2.0
alembic==1.7.7
python-jose[cryptography]
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
pydantic-settings==2.5.2